Remove unnecessary VIP html attribute replacement
It turned out that all the VIP uses of the button just called render button directly passing attributes. As these won't pass inner block content they are treated as deprecated buttons for rendering. This means we won't be getting submitButtonAttributes while rendering the newest versions of the block added through the editor.